feat: 🎸 A new Brewkit structure
New structure for Brewkit revolving around lerna.js a monorepo manager.
This repo is independent as that of an actual npm package, and is
organized in manner to link versions of multiple npm packages (eg.
components, and themes for now), do a joint build, and provide dev tools
for all this packages on once place. Storybook itself will be used on
/docs, which is in fact a main package that will be used to serve
contents from Next.js framework.